chore: pr feedback, updates to pagination, add linting/tsc command
