+++ packages/react/src/jsx/components/UqPagination/UqPaginationItem.tsx
@@ -1,12 +1,7 @@
-import { FC, MouseEventHandler } from "react";
+import type { FC, MouseEventHandler } from "react";
 import classNames from "classnames";
 import { UqIcon } from "../UqIcon/UqIcon.tsx";
-
-export enum UqPaginationItemType {
-  Previous = "previous",
-  Next = "next",
-  Ellipsis = "ellipsis",
-}
+import { UqPaginationItemType } from "../../enum";
 
 export interface UqPaginationItemProps {
   As?: "button" | "a";
@@ -14,7 +9,7 @@ export interface UqPaginationItemProps {
   number?: number;
   current?: boolean;
   href?: string;
-  value: string;
+  value?: string;
   onClick?: MouseEventHandler<HTMLButtonElement>;
 }
 
@@ -26,32 +21,35 @@ export const UqPaginationItem: FC<UqPaginationItemProps> = ({
   href,
   value,
   onClick,
-}) => (
-  <li>
-    <As
-      className={classNames("uq-pagination__item", {
-        [`uq-pagination__${type}`]: type,
-        "uq-pagination__current": current,
-      })}
-      aria-current={current && "page"}
-      onClick={onClick}
-      href={href}
-      value={value}
-    >
-      <span className="visually-hidden">
-        {!type && "Page"}
-        {type === UqPaginationItemType.Previous && "Previous page"}
-        {type === UqPaginationItemType.Next && "Previous page"}
-        {type === UqPaginationItemType.Ellipsis && `Page ${number}`}
-      </span>
-      {!type && `${number}`}
-      {type === UqPaginationItemType.Previous && (
-        <UqIcon name={"standard--chevron-left"} />
-      )}
-      {type === UqPaginationItemType.Next && (
-        <UqIcon name={"standard--chevron-right"} />
-      )}
-      {type === UqPaginationItemType.Ellipsis && "…"}
-    </As>
-  </li>
-);
+}) => {
+  return (
+    <li className="uq-pagination__list-item">
+      <As
+        className={classNames("uq-pagination__item", {
+          [`uq-pagination__${type}`]: type,
+          "uq-pagination__current": current,
+        })}
+        aria-current={current && "page"}
+        // @ts-expect-error Dynamic elements expect different attributes (a and button).
+        onClick={onClick}
+        href={href}
+        value={value}
+      >
+        <span className="visually-hidden">
+          {!type && "Page"}
+          {type === UqPaginationItemType.Previous && "Previous page"}
+          {type === UqPaginationItemType.Next && "Previous page"}
+          {type === UqPaginationItemType.Ellipsis && `Skip to page ${number}`}
+        </span>
+        {!type && `${number}`}
+        {type === UqPaginationItemType.Previous && (
+          <UqIcon name={"standard--chevron-left"} />
+        )}
+        {type === UqPaginationItemType.Next && (
+          <UqIcon name={"standard--chevron-right"} />
+        )}
+        {type === UqPaginationItemType.Ellipsis && "…"}
+      </As>
+    </li>
+  );
+};
