+++ packages/react/src/jsx/enum/index.ts
@@ -1,0 +1,42 @@
+export const enum UqCardVariant {
+  text = "text",
+  image = "image",
+  icon = "icon",
+  iconNav = "icon-nav",
+}
+
+export enum UqCardGridTarget {
+  x1 = "1x",
+  x2 = "2x",
+  x3 = "3x",
+  x4 = "4x",
+}
+
+export const enum UqIconVariant {
+  light = "light",
+  text = "text",
+}
+
+export const enum UqPaginationItemType {
+  Previous = "previous",
+  Next = "next",
+  Ellipsis = "ellipsis",
+}
+
+export const enum UqPaneVariant {
+  text = "text",
+  image = "image",
+  video = "video",
+  icon = "icon",
+  iconLarge = "icon-large",
+  textShaded = "text-shaded",
+  iconShaded = "icon-shaded",
+}
+
+export const enum UqStoryCardVariant {
+  story = "story",
+  news = "news",
+  event = "event",
+  landscape = "landscape",
+  feature = "feature",
+}
