+++ packages/react/src/jsx/layout/UqContainer/UqContainer.tsx
@@ -1,4 +1,4 @@
-import { FC, PropsWithChildren } from "react";
+import type { FC, PropsWithChildren } from "react";
 
 export interface UqContainerProps extends PropsWithChildren {}
