+++ packages/react/tsconfig.json
@@ -16,7 +16,8 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
-    "declaration": true
+    "declaration": true,
+    "verbatimModuleSyntax": true
   },
   "include": ["src"]
 }
