+++ packages/storybook-react/stories/Components/UqCard/UqCard.stories.jsx
@@ -1,6 +1,9 @@
 import { reactCardDecorator } from "../../../lib/decorators";
-import { UqCard as Component, UqIcon } from "@uqds/react/src/jsx/main";
-import { UqCardVariant } from "@uqds/react/src/jsx/components/UqCard/UqCard";
+import {
+  UqCard as Component,
+  UqCardVariant,
+  UqIcon,
+} from "@uqds/react/src/jsx/main";
 
 export default {
   title: "Components/UqCard",
