+++ packages/storybook-react/stories/Components/UqPagination/UqPagination.mdx
@@ -1,0 +1,26 @@
+import { Meta, Canvas, Controls, Source } from "@storybook/blocks";
+import * as UqPaginationStories from "./UqPagination.stories";
+
+<Meta of={UqPaginationStories} />
+
+# UqPagination
+
+The pagination component.
+
+## Installation
+
+```bash
+npm i @uqds/react
+```
+
+## Usage
+
+```js
+import { UqPagination } from "@uqds/react";
+```
+
+<Canvas of={UqPaginationStories.UqPagination} sourceState="none" />
+
+<Controls of={UqPaginationStories.UqPagination} />
+
+<Source of={UqPaginationStories.UqPagination} />
